Some UI improvements, Added Flyway migration library, Entities has been reworked, We use Decimal in the sql and BigDecimal in the backend for the price value, Some configuration (Hibernate table creation is now disabled), CSRF is now enabled for PROD environment, Added some basic tests for users and services

+++ src/main/java/com/martinatanasov/computerstore/controllers/ProductController.java
@@ -59,7 +59,7 @@ public class ProductController {
                       @RequestParam(required = false, defaultValue = "1") Integer pageNumber,
                       @RequestParam(required = false, defaultValue = "3") Integer pageSize,
                       @RequestParam(required = false, defaultValue = "asc") String sortValue){
-        Page<Product> products = productService.findAllByCategoryId(1L, pageNumber, pageSize, sortValue);
+        Page<Product> products = productService.findAllByCategoryId((short) 1, pageNumber, pageSize, sortValue);
         if(!products.isEmpty()){
             Page<StoreItem> dtoProducts = productConverter.convertToStoreItems(products);
             if(pageNumber > dtoProducts.getTotalElements() || pageNumber < 1){
@@ -81,7 +81,7 @@ public class ProductController {
                            @RequestParam(required = false, defaultValue = "1") Integer pageNumber,
                            @RequestParam(required = false, defaultValue = "3") Integer pageSize,
                            @RequestParam(required = false, defaultValue = "asc") String sortValue){
-        Page<Product> products = productService.findAllByCategoryId(2L, pageNumber, pageSize, sortValue);
+        Page<Product> products = productService.findAllByCategoryId((short) 2, pageNumber, pageSize, sortValue);
         if(!products.isEmpty()){
             Page<StoreItem> dtoProducts = productConverter.convertToStoreItems(products);
             if(pageNumber > dtoProducts.getTotalElements() || pageNumber < 1){
@@ -103,7 +103,7 @@ public class ProductController {
                              @RequestParam(required = false, defaultValue = "1") Integer pageNumber,
                              @RequestParam(required = false, defaultValue = "3") Integer pageSize,
                              @RequestParam(required = false, defaultValue = "asc") String sortValue){
-        Page<Product> products = productService.findAllByCategoryId(3L, pageNumber, pageSize, sortValue);
+        Page<Product> products = productService.findAllByCategoryId((short) 3, pageNumber, pageSize, sortValue);
         if(!products.isEmpty()){
             Page<StoreItem> dtoProducts = productConverter.convertToStoreItems(products);
             if(pageNumber > dtoProducts.getTotalElements() || pageNumber < 1){
@@ -121,7 +121,7 @@ public class ProductController {
     }
 
     @GetMapping("/{productId}")
-    public String itemReview(@PathVariable(value = "productId") Long productId,
+    public String itemReview(@PathVariable(value = "productId") Integer productId,
                              Model model){
         Optional<Product> product = productService.getProductById(productId);
         product.ifPresent(value -> model.addAttribute("product", productConverter.convertToSingleItem(value)));
